+++ src/Products/Sessions/BrowserIdManager.py
@@ -86,8 +86,7 @@ class BrowserIdManager(Item, Persistent, Implicit, RoleManager, Owned, Tabs):
 
     security = ClassSecurityInfo()
     security.declareObjectPublic()
-    ok = {'meta_type':1, 'id':1, 'title': 1, 'icon':1,
-          'bobobase_modification_time':1, 'title_or_id':1 }
+    ok = {'meta_type': 1, 'id': 1, 'title': 1, 'icon': 1, 'title_or_id': 1}
     security.setDefaultAccess(ok)
     security.setPermissionDefault(MGMT_SCREEN_PERM, ['Manager'])
     security.setPermissionDefault(ACCESS_CONTENTS_PERM,['Manager','Anonymous'])