+++ src/Products/Sessions/SessionDataManager.py
@@ -79,8 +79,7 @@ class SessionDataManager(Item, Implicit, Persistent, RoleManager, Owned, Tabs):
     security = ClassSecurityInfo()
     security.declareObjectPublic()
 
-    ok = {'meta_type':1, 'id':1, 'title': 1, 'icon':1,
-          'bobobase_modification_time':1, 'title_or_id':1 }
+    ok = {'meta_type': 1, 'id': 1, 'title': 1, 'icon': 1, 'title_or_id': 1}
     security.setDefaultAccess(ok)
     security.setPermissionDefault(CHANGE_DATAMGR_PERM, ['Manager'])
     security.setPermissionDefault(MGMT_SCREEN_PERM, ['Manager'])